+++ restapi/src/routes/ProductRoutes.ts
@@ -19,5 +19,7 @@ router.get('/product/getByCategoria/:categoria', ProductController.getProductsBy
 router.get('/product/getByPrecio/:precio', ProductController.getProductByPrice);
 // Productos añadir con Post
 router.post('/product/addPost', ProductController.addProductPost);
+
+router.get('/product/getByCodigo/:codigo', ProductController.getProductoByCodigo);
 module.exports = router;
 export default router;
