+++ webapp/src/api/ApiProducts.ts
@@ -41,6 +41,6 @@ export async function addProduct(url: string, nombre: string, descripcion: strin
 
 export async function getProductByCode (code: string): Promise<Product> {
     const apiEndPoint= process.env.REACT_APP_API_URI || 'http://localhost:5000'
-    let response = await fetch(apiEndPoint + "/product/getByCode/" + code);
+    let response = await fetch(apiEndPoint + "/product/getByCodigo/" + code, {method: "GET"});
     return response.json();
 }
